+++ HISTORY.md
@@ -4,6 +4,7 @@
 - Implemented access lookup and improved attribute handling
 - Normalized formatting on the low-level client modules
 - Added embedded attribute handling to Person entities
+- DN Lookup for person entities
 
 ## 0.3.0
 - Advanced options for suspend/restore person